+++ game/Lmd_Prof_Core.c
@@ -1150,7 +1150,7 @@ void Experience_Level_Up(gentity_t *ent)
 		else if (playerLevel == 100)PlayerAcc_SetLootboxes(ent, PlayerAcc_GetLootboxes(ent) + 20);
 		else if (playerLevel == 120) {
 			PlayerAcc_SetLootboxes(ent, PlayerAcc_GetLootboxes(ent) + 50);
-			Disp(ent, "^5You've got a special bonus for reaching ^3level Cap!\n ^2Don't hesitate to try a ^3New Game Plus ^2mode.");
+			Disp(ent, "^5You've got a special bonus for reaching ^3level Cap!\n^2Don't hesitate to try a ^3New Game Plus ^2mode.");
 		}
 		/////
 
@@ -1162,9 +1162,8 @@ void Experience_Level_Up(gentity_t *ent)
 
 	   cost = Professions_LevelCost_EXP(prof, playerLevel);
 	   Disp(ent, va("^5%i ^3/ ^2%i ^3EXP\n", resEXP, cost));      
-	   char *msg_other = va("^1%s ^3becomes more powerful. One's level is ^1%i^3 now.\n.", PlayerAcc_GetName(ent), playerLevel);
+	   char *msg_other = va("\n^1%s ^3becomes more powerful. One's level is ^1%i^3 now.\n.", PlayerAcc_GetName(ent), playerLevel);
 	 char *msg = va("^5Congratulation! Level Increased!\n^3Your level is ^2%i^3.\n^3%i skill points recived.", playerLevel, NewSkillPoints_value);
-		   
 	   Disp(ent, msg);
 	   trap_SendServerCommand(-1, va("print \"%s\"", msg_other));
 	   trap_SendServerCommand(ent->s.number, va("cp \"%s\"", msg));